Remove SearchResults and use SearchResultGroup directly
